moved our rjsf form to component so extensions can also use it w/ burnettk
